+++ src/MiloneTranslation/TranslationApi.fs
@@ -7,6 +7,7 @@ open MiloneShared.SharedTypes
 open Std.StdMap
 open MiloneTranslationTypes.HirTypes
 open MiloneTranslationTypes.TranslationApiTypes
+open MiloneTranslation.Cir
 
 module S = Std.StdString
 module AutoBoxing = MiloneTranslation.AutoBoxing
@@ -22,7 +23,11 @@ module MonoTy = MiloneTranslation.MonoTy
 module RecordRes = MiloneTranslation.RecordRes
 module TailRecOptimizing = MiloneTranslation.TailRecOptimizing
 
-let private codeGenHir (writeLog: WriteLogFun) (modules: HProgram, hirCtx: HirCtx) : (DocId * CCode) list =
+let private codeGenHir
+  entrypointName
+  (writeLog: WriteLogFun)
+  (modules: HProgram, hirCtx: HirCtx)
+  : (DocId * CCode) list * ExportName list =
   writeLog "RecordRes"
   let modules, hirCtx = RecordRes.recordRes (modules, hirCtx)
 
@@ -80,7 +85,9 @@ let private codeGenHir (writeLog: WriteLogFun) (modules: HProgram, hirCtx: HirCt
   let modules, mirCtx = MirGen.mirify (modules, hirCtx)
 
   writeLog "CirGen"
-  let modules = CirGen.genCir (modules, mirCtx)
+
+  let modules =
+    CirGen.genCir entrypointName (modules, mirCtx)
 
   writeLog "CirDump"
 
@@ -88,6 +95,13 @@ let private codeGenHir (writeLog: WriteLogFun) (modules: HProgram, hirCtx: HirCt
     modules
     |> List.map (fun (docId, cir) -> docId, CirDump.cirDump cir)
 
-  cFiles
+  let exportNames =
+    modules |> List.collect (fun (_, decls) ->
+      decls |> List.choose (fun decl ->
+        match decl with
+        | CFunDecl (ident, _, _, _) -> Some ident
+        | _ -> None ))
+
+  cFiles, exportNames
 
 let newTranslationApi () : TranslationApi = { CodeGenHir = codeGenHir }